examples/integrations/requirements.txt
aiofiles==0.7.0
anyio==3.3.3
asgiref==3.4.1
Authlib==0.15.4
caio==0.9.3
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
fastapi==0.68.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pavlok==0.1.3
pyclickup==0.1.4
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
